+++ cosmic/sample/initialcmctable.py
@@ -62,10 +62,22 @@ class InitialCMCTable(pd.DataFrame):
     mass_of_cluster = None
     virial_radius = None
     tidal_radius = None
+    central_bh = 0.0
+    scale_with_central_bh = False
+
+    def ScaleCentralBHMass(self, Mtotal):
+        """Rescale the central BH mass; needed since this is a class attribute
+            Parameters
+            ----------
+
+            Mtotal : float
+                total mass of the cluster
+        """
+        self.central_bh /= Mtotal
 
     @classmethod
-    def ScaleToNBodyUnits(cls, Singles, Binaries, virial_radius=1):
-        r"""Rescale the single masses, radii, and velocities into N-body units
+    def ScaleToNBodyUnits(cls, Singles, Binaries, virial_radius=1, central_bh=0, scale_with_central_bh=False):
+        """Rescale the single masses, radii, and velocities into N-body units
            i.e. \sum m = M = 1
                  Kinetic Energy   = 0.25
                  Potential Energy = -0.5
@@ -94,6 +106,10 @@ class InitialCMCTable(pd.DataFrame):
         Binaries["m1"] /= M_total
         Binaries["m2"] /= M_total
 
+        # Note if there's an central BH, we want to add it
+        # to the mass of the already normalized stars (i.e. M_tot = 1+BH)
+        Singles.ScaleCentralBHMass(M_total)
+
         # Take the radii, and offset by one
         radius = np.array(Singles["r"])
         radius_p1 = np.append(radius[1:], [1e100])
@@ -107,11 +123,15 @@ class InitialCMCTable(pd.DataFrame):
         # Then compute the total kinetic and potential energy
         # There's probably a cleaner way to do the PE (this is a one-line version
         #  of the for loop we use in CMC; vectorized and pythonic, but sloppy)
-        KE = 0.5 * sum(mass * (vr ** 2 + vt ** 2))
-        PE = 0.5 * sum(
+        KE = 0.5 * np.sum(mass * (vr ** 2 + vt ** 2))
+        PE = 0.5 * np.sum(
             mass[::-1]
             * np.cumsum((cumul_mass * (1.0 / radius - 1.0 / radius_p1))[::-1])
         )
+
+        # If scaling with central BH, add the potential from it
+        if scale_with_central_bh:
+            PE += np.sum(Singles.central_bh*mass / radius) 
 
         # Compute the position and velocity scalings
         rfac = 2 * PE
@@ -276,18 +296,20 @@ class InitialCMCTable(pd.DataFrame):
         virial_radius = kwargs.pop('virial_radius',Singles.virial_radius)
         tidal_radius = kwargs.pop('tidal_radius',Singles.tidal_radius)
         metallicity = kwargs.pop('metallicity',Singles.metallicity)
+        central_bh = kwargs.pop('central_bh',Singles.central_bh)
+        scale_with_central_bh = kwargs.pop('scale_with_central_bh',Singles.scale_with_central_bh)
 
         # If a user has not already scaled the units of the Singles and Binaries tables,
         # and the attribute mass_of_cluster is None, then
         # we can calculate it now
         if (not Singles.scaled_to_nbody_units) and (Singles.mass_of_cluster is None):
-            Singles.mass_of_cluster = np.sum(Singles["m"])
+            Singles.mass_of_cluster = np.sum(Singles["m"]) + central_bh
             InitialCMCTable.ScaleToNBodyUnits(
-                Singles, Binaries, virial_radius=virial_radius
+                Singles, Binaries, virial_radius=virial_radius, central_bh=central_bh, scale_with_central_bh=scale_with_central_bh
             )
         elif (not Singles.scaled_to_nbody_units) and (Singles.mass_of_cluster is not None):
             InitialCMCTable.ScaleToNBodyUnits(
-                Singles, Binaries, virial_radius=virial_radius
+                Singles, Binaries, virial_radius=virial_radius, central_bh=central_bh, scale_with_central_bh=scale_with_central_bh
             )
         elif (Singles.scaled_to_nbody_units) and (Singles.mass_of_cluster is None):
             # we cannot get the pre-scaled mass of the cluster