+++ cosmic/sample/sampler/cmc.py
@@ -38,7 +38,7 @@ __all__ = ["get_cmc_sampler", "CMCSample"]
 def get_cmc_sampler(
     cluster_profile, primary_model, ecc_model, porb_model, binfrac_model, met, size, **kwargs
 ):
-    """Generates an initial binary sample according to user specified models
+    """Generates an initial cluster sample according to user specified models
 
     Parameters
     ----------
@@ -122,6 +122,14 @@ def get_cmc_sampler(
         the initial tidal radius of the cluster, in units of the virial_radius
         Default -- 1e6 rvir
 
+    central_bh : `float`
+        Put a central massive black hole in the cluster
+        Default -- 0 MSUN
+        
+    scale_with_central_bh : `bool`
+        If True, then the potential from the central_bh is included when scaling the radii.
+        Default -- False
+
     seed : `float`
         seed to the random number generator, for reproducability
 
@@ -133,7 +141,6 @@ def get_cmc_sampler(
     Binaries: `pandas.DataFrame`
         DataFrame of Single objects in the format of the InitialCMCTable
     """
-    
     initconditions = CMCSample()
 
     # if RNG seed is provided, then use it globally
@@ -222,10 +229,12 @@ def get_cmc_sampler(
     )
 
     singles_table.metallicity = met
-    singles_table.mass_of_cluster = np.sum(singles_table["m"])
     binaries_table.metallicity = met
     singles_table.virial_radius = kwargs.get("virial_radius",1) 
-    singles_table.tidal_radius = kwargs.get("tidal_radius",1e6) 
+    singles_table.tidal_radius = kwargs.get("tidal_radius",1e13) 
+    singles_table.central_bh = kwargs.get("central_bh",0)
+    singles_table.scale_with_central_bh = kwargs.get("scale_with_central_bh",False)
+    singles_table.mass_of_cluster = np.sum(singles_table["m"]) + singles_table.central_bh
 
     return singles_table, binaries_table
 
@@ -240,7 +249,8 @@ register_sampler(
 def get_cmc_point_mass_sampler(
     cluster_profile, size, **kwargs
 ):
-    """Generates an CMC cluster model according  
+    """Generates an CMC cluster model according to user-specified model.
+    Note here that masses will all be unity (with total cluster normalized accordingly)
 
     Parameters
     ----------
@@ -266,6 +276,29 @@ def get_cmc_point_mass_sampler(
     size : `int`
         Size of the population to sample
 
+    Optional Parameters
+    -------------------
+    virial_radius : `float`
+        the initial virial radius of the cluster, in parsecs
+        Default -- 1 pc
+
+    tidal_radius : `float`
+        the initial tidal radius of the cluster, in units of the virial_radius
+        Default -- 1e6 rvir
+
+    central_bh : `float`
+        Put a central massive black hole in the cluster.  Note here units are 
+        in code units where every particle has mass 1 (before normalization), e.g. if you
+        want a cluster with 10000 stars and a BH that is 10% of the total mass, central_bh=1000
+        Default -- 0
+        
+    scale_with_central_bh : `bool`
+        If True, then the potential from the central_bh is included when scaling the radii.
+        Default -- False
+
+    seed : `float`
+        seed to the random number generator, for reproducability
+
     Returns
     -------
     Singles: `pandas.DataFrame`
@@ -274,8 +307,12 @@ def get_cmc_point_mass_sampler(
         DataFrame of Single objects in the format of the InitialCMCTable
 
     """
-
     initconditions = CMCSample()
+
+    # if RNG seed is provided, then use it globally
+    rng_seed = kwargs.pop("seed", 0)
+    if rng_seed != 0:
+        np.random.seed(rng_seed)
 
     # get radii, radial and transverse velocities
     r, vr, vt = initconditions.set_r_vr_vt(cluster_profile, N=size, **kwargs)
@@ -288,17 +325,29 @@ def get_cmc_point_mass_sampler(
     singles_table = InitialCMCTable.InitialCMCSingles(
         single_ids + 1, initconditions.set_kstar(mass1), mass1, Reff, r, vr, vt, binind
     )
-    # Already scaled from the IC generators
-    singles_table.scaled_to_nbody_units = True
 
     # We assume no binaries for the point mass models
     binaries_table = InitialCMCTable.InitialCMCBinaries(1,1,0,0,0,1,0,0,0,0,0)
 
     singles_table.metallicity = 0.02
-    singles_table.mass_of_cluster = np.sum(singles_table["m"])*size
     binaries_table.metallicity = 0.02
-    singles_table.virial_radius = kwargs.get("virial_radius",1) 
-    singles_table.tidal_radius = kwargs.get("tidal_radius",1e6) 
+    singles_table.virial_radius = kwargs.get("virial_radius",1)
+    singles_table.tidal_radius = kwargs.get("tidal_radius",1e13)
+    singles_table.central_bh = kwargs.get("central_bh",0)
+    singles_table.scale_with_central_bh = kwargs.get("scale_with_central_bh",False)
+    singles_table.mass_of_cluster = np.sum(singles_table["m"])*size + singles_table.central_bh
+
+    # Already scaled from the IC generators (unless we've added a central BH)
+    if singles_table.central_bh != 0:
+        singles_table.scaled_to_nbody_units = False
+    else:
+        singles_table.scaled_to_nbody_units = True
+
+    # Already scaled from the IC generators (unless we've added a central BH)
+    if singles_table.central_bh != 0:
+        singles_table.scaled_to_nbody_units = False 
+    else:
+        singles_table.scaled_to_nbody_units = True
    
     return singles_table, binaries_table