+++ changelog.md
@@ -38,3 +38,4 @@ See the discussed changes in our previous releases here: https://github.com/COSM
  - Bug fixes: `timestep_conditions` for `bcm` arrays now raise errors for invalid columns instead of failing silently
  - Add `teff_1` and `teff_2` as variables that can be used to set `timestep_conditions`
  - Add in `-1` option to turn off Magnetic Braking in htmb 
+ - Added `central_bh` and `scale_with_central_bh` as options to the CMC sampler, in order to add central massive black holes to CMC initial conditions